Split giant ByteCodeEngine tests into smaller sets
