+++ Assets/ulox/Tests/ClassTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 530949fa27f3b274d94a25ce7578e06a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
