+++ Assets/ulox/Tests/DynamicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93d1e8d7ce6fd6345895dae724dd80f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
